requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
python-dotenv>=1.0
xhtml2pdf>=0.2.0
reportlab>=4.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
python-docx>=1.1.0
